Added a new Project "Challenges".
My goal is to restructure my `Advent of Code` challenges into an organized project instead of one project per challenge. Therefor the user will be able to choose which challenge to run and it will also automatically be timed and the results shown on the screen. If I feel like it, I would consider adding a visualisation of challenges, like showing mazes or whatever the challenges has potential to show to the user.